+++ packages/extension-ui/src/Popup/Accounts/Account.test.tsx
@@ -33,19 +33,19 @@ describe('Account component', () => {
     wrapper = mountAccountComponent({ isExternal: false });
     wrapper.find('Details').simulate('click');
 
-    expect(wrapper.find('MenuItem').length).toBe(4);
-    expect(wrapper.find('MenuItem').at(0).text()).toBe('Rename');
-    expect(wrapper.find('MenuItem').at(1).text()).toBe('Derive New Account');
-    expect(wrapper.find('MenuItem').at(2).text()).toBe('Export Account');
-    expect(wrapper.find('MenuItem').at(3).text()).toBe('Forget Account');
+    expect(wrapper.find('a.menuItem').length).toBe(4);
+    expect(wrapper.find('a.menuItem').at(0).text()).toBe('Rename');
+    expect(wrapper.find('a.menuItem').at(1).text()).toBe('Derive New Account');
+    expect(wrapper.find('a.menuItem').at(2).text()).toBe('Export Account');
+    expect(wrapper.find('a.menuItem').at(3).text()).toBe('Forget Account');
   });
 
   it('does not show Export option if account is external', () => {
     wrapper = mountAccountComponent({ isExternal: true });
     wrapper.find('Details').simulate('click');
 
-    expect(wrapper.find('MenuItem').length).toBe(2);
-    expect(wrapper.find('MenuItem').at(0).text()).toBe('Rename');
-    expect(wrapper.find('MenuItem').at(1).text()).toBe('Forget Account');
+    expect(wrapper.find('a.menuItem').length).toBe(2);
+    expect(wrapper.find('a.menuItem').at(0).text()).toBe('Rename');
+    expect(wrapper.find('a.menuItem').at(1).text()).toBe('Forget Account');
   });
 });